bots/requirements.txt
pytrends==4.9.2
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
lxml>=5.0.0
anthropic>=0.40.0
urllib3<2
pandas>=2.0.0
